backend/requirements.txt
fastapi==0.114.2
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.36
psycopg2-binary==2.9.9
python-jose==3.3.0
passlib==1.7.4
argon2-cffi==23.1.0
pydantic-settings==2.4.0
email-validator==2.2.0
alembic==1.13.2
yookassa==3.3.0
minio==7.2.3



# Monitoring
prometheus-fastapi-instrumentator==6.1.0
prometheus-client==0.20.0
